+++ src/common/utilities.hpp
@@ -292,7 +292,7 @@ namespace rathena {
 		* @param padding: Padding character
 		* @param num: Maximum length of padding
 		*/
-		void string_left_pad_inplace(std::string& str, char padding, int num);
+		void string_left_pad_inplace(std::string& str, char padding, size_t num);
 
 		/**
 		* Pad string with arbitrary character
@@ -302,7 +302,7 @@ namespace rathena {
 		*
 		* @return A copy of original string with padding added
 		*/
-		std::string string_left_pad(const std::string& original, char padding, int num);
+		std::string string_left_pad(const std::string& original, char padding, size_t num);
 
 		/**
 		* Encode base10 number to base62. Originally by lututui
